+++ sources/gookto_io/intersection/port.cpp
@@ -3,7 +3,7 @@
 
 #include <gookto_io/intersection/port.hpp>
 
-port::port(long int id, int dir, int type) :
+port::port(int id, int dir, int type) :
 linkId(id), direction(dir), type(type)
 {};
